+++ changelog.md
@@ -5,6 +5,7 @@
 - Make all biomes in `#minecraft:is_nether` tag marked as having HELLISH climate (#65)
 - The Phantasmal line (Ended, Spectral, Phantasmal) is now actually nocturnal and only works during the night time (unless given the NEVER_SLEEPS allele).
 - Added Creative Frames for debugging Forestry mutations.
+- Hive frames can now stack up to 64.
 
 ## Forestry 1.0.5
 - Fixed wild beehives not having particle effects